# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecureShare Authorization Reference API",
    "summary": "Multi-tenant document sharing with JWT, OpenFGA, Macaroons, PostgreSQL, and audit logging.",
    "description": "\nSecureShare is a production-style reference API for secure multi-tenant document sharing.\n\nCore security properties:\n\n* JWTs carry identity only, never document permissions.\n* FastAPI routers and dependencies act as policy enforcement points.\n* OpenFGA is the relationship-based policy decision point for document access.\n* Macaroon-style delegated links provide temporary, attenuated read access.\n* Every allow and deny authorization decision is written to the audit log.\n\nDemo users all use password `password123`:\n\n* `alice@example.com` owns the Acme document.\n* `bob@example.com` can view the Acme document.\n* `eve@example.net` owns the Globex document in another tenant.\n",
    "contact": {
      "name": "SecureShare maintainers"
    },
    "license": {
      "name": "Reference implementation"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create a document",
        "description": "Creates a document in the caller's tenant only after the shared authorization service verifies live tenant membership, then stores application data in PostgreSQL and writes OpenFGA owner and tenant-parent relationships.",
        "operationId": "create_document_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Document created and owner relationship written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "Caller is not an active member of the tenant."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Read a document",
        "description": "Reads a document only after the shared authorization service verifies tenant boundary and OpenFGA returns `can_read` for the caller.",
        "operationId": "read_document_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "OpenFGA denied access or the document is cross-tenant."
          },
          "404": {
            "description": "Document not found after authorization succeeds."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "documents"
        ],
        "summary": "Edit a document",
        "description": "Updates a document only after the shared authorization service verifies tenant boundary and OpenFGA returns `can_edit` for the caller.",
        "operationId": "update_document_documents__document_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "OpenFGA denied edit access or the document is cross-tenant."
          },
          "404": {
            "description": "Document not found after authorization succeeds."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/relationships": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Inspect document relationships",
        "description": "Returns OpenFGA relationships for a document. This admin endpoint requires tenant-admin access and only reveals relationships for documents inside the caller's tenant.",
        "operationId": "inspect_relationships_documents__document_id__relationships_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document relationships returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipInspection"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "Caller is not allowed to inspect relationships."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/delegated/documents/{document_id}": {
      "get": {
        "tags": [
          "delegated"
        ],
        "summary": "Read a document using a delegated token",
        "description": "Reads a document with an `x-delegation-token` Macaroon. The endpoint validates caveats and performs a live OpenFGA `can_read` check for the issuing user before returning data.",
        "operationId": "read_delegated_document_delegated_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "x-delegation-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Delegation-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document returned through delegated access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "403": {
            "description": "Delegated token is expired, malformed, caveat-violating, or issuer access was revoked."
          },
          "404": {
            "description": "Document not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "DocumentCreate",
        "examples": [
          {
            "body": "Confidential launch notes for the Acme tenant.",
            "title": "Quarterly Launch Plan"
          }
        ]
      },
      "DocumentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "owner_id",
          "title",
          "body",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead"
      },
      "DocumentUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "type": "object",
        "title": "DocumentUpdate",
        "examples": [
          {
            "body": "Updated document body.",
            "title": "Updated Quarterly Launch Plan"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationshipInspection": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipRead"
            },
            "type": "array",
            "title": "Relationships"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "relationships"
        ],
        "title": "RelationshipInspection"
      },
      "RelationshipRead": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User"
          },
          "relation": {
            "type": "string",
            "title": "Relation"
          },
          "object": {
            "type": "string",
            "title": "Object"
          }
        },
        "type": "object",
        "required": [
          "user",
          "relation",
          "object"
        ],
        "title": "RelationshipRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}